rename:: timetable/FriendTimetable -> timetable/Friend - 기존 Friend 관련 컴포넌트는 시간표가 아닌 친구 기능에만 방점이 있어 재명명
