+++ src/components/timetable/Friend/FriendCard.tsx
