+++ src/components/timetable/Friend/FriendsList.tsx
@@ -3,26 +3,6 @@ import { Link } from 'react-router-dom'
 
 import { useGetFriendList } from '@/api/hooks/friends'
 
-const PlusFriend = css({
-  w: '100%',
-  bgColor: 'bg.gray',
-  py: 3.5,
-  rounded: 10,
-  border: '1px {colors.lightGray.1} solid',
-  color: 'lightGray.1',
-  fontSize: 18,
-  fontWeight: 500,
-  cursor: 'pointer',
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  transition: 'border 0.256s, color 0.256s',
-  _hover: {
-    borderColor: 'darkGray.2',
-    color: 'darkGray.2',
-  },
-})
-
 const FriendBlock = css.raw({
   h: 11,
   w: 30,
@@ -41,14 +21,41 @@ const FriendBlock = css.raw({
   },
 })
 
+export const FriendPageBtn = () => {
+  return (
+    <Link
+      className={css({
+        w: 47,
+        bgColor: 'bg.gray',
+        py: 3.5,
+        rounded: 10,
+        border: '1px {colors.lightGray.1} solid',
+        color: 'lightGray.1',
+        fontSize: 18,
+        fontWeight: 500,
+        cursor: 'pointer',
+        display: 'flex',
+        justifyContent: 'center',
+        alignItems: 'center',
+        transition: 'border 0.256s, color 0.256s',
+        _hover: {
+          borderColor: 'darkGray.2',
+          color: 'darkGray.2',
+        },
+      })}
+      to={'/timetable/friend'}
+    >
+      Add friends
+    </Link>
+  )
+}
+
 const FriendsList = () => {
   const { data } = useGetFriendList({ keyword: null })
 
   return (
-    <div className={css({ w: 47, display: 'flex', flexDir: 'column', gap: 3.5, alignItems: 'center' })}>
-      <Link className={PlusFriend} to={'/timetable/friend'}>
-        Add friends
-      </Link>
+    <div className={css({ display: 'flex', flexDir: 'column', gap: 3.5, alignItems: 'center' })}>
+      <FriendPageBtn />
       <div
         className={css({
           display: 'flex',
